20/03/2025 Update Sql kiểu dữ liệu của luongCB của bảng CHUCVU Thêm các class cho bảng BANGCHAMCONG

+++ src/BUS/NhanVienBUS.java
@@ -64,7 +64,7 @@ public class NhanVienBUS {
 		return roleName;
 	}
 	
-	public float getBaseSalaryByRoleID(String maCV) {
+	public Double getBaseSalaryByRoleID(String maCV) {
 		return nvDAO.getBaseSalaryByRoleID(maCV);
 	}
